Add find news by keyword & remove error from connection pool

+++ NewsPortal/src/main/java/tcejorptset/servl/controller/impl/DoDeleteNews.java
@@ -21,17 +21,18 @@ public class DoDeleteNews implements Command {
 		
 		String [] newsId = request.getParameterValues(AttributeParamName.JSP_NEWS_ID_PARAM);
 			try {
+				if(newsId != null) {
 				int [] id = Stream.of(newsId)
 						  .mapToInt(Integer::parseInt)
 						  .toArray();
 				newsService.delete(id);
+				}
 				response.sendRedirect("controller?command=go_to_news_list");
 			} catch (ServiceException e) {
 				request.getSession().setAttribute (AttributeParamName.JSP_ERROR_CODE_ATTRIBUTE, ErrorCode.DELETE_NEWS.getCode());
 				response.sendRedirect("error?command=go_to_news_list");
 				e.printStackTrace();
 			}
-				
 		}
 	}
